Financial-tool-main/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pymongo==4.12.1
dnspython==2.7.0
pyjwt==2.10.1
bcrypt==4.3.0
openpyxl==3.1.5
python-dotenv==1.1.0
